% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all.int}
\alias{all.int}
\title{Check if all are integers}
\usage{
\method{all}{int}(x, ...)
}
\arguments{
\item{x, ...}{objects to be checked}
}
\value{
boolean value
}
\description{
Using funcion is.int from this package, checks if all given elements are actual integers.
}
