% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amicable.R
\name{amicable}
\alias{amicable}
\title{Check if two numbers are amicable.}
\usage{
amicable(x, y)
}
\arguments{
\item{x}{The first number.}

\item{y}{The second number.}
}
\value{
A logical TRUE or FALSE.
}
\description{
Does not have any dependencies, but for assignment purposes will suggest dplyr to be installed.
}
\details{
Two distinct positive integers are amicable if their sums of their proper divisors are equal.
}
\examples{
amicable(220, 284)
amicable(10, 12)
}
