% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars}
\alias{cars}
\title{Speed and Stopping Distances of Cars.}
\format{The cars data frame contains 50 observations on 2 variables.
\describe{
  \item{speed}{Speed, in mph}
  \item{dist}{Weight, in ft}
}}
\source{
cars # from R Dataset Package
}
\usage{
cars
}
\description{
The data give the speed of cars and the distances taken to stop.
}
\keyword{datasets}
