% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coprime}
\alias{coprime}
\title{Checks coprimity of numbers}
\usage{
coprime(a, b)
}
\arguments{
\item{a, b}{given numbers}
}
\value{
boolean value
}
\description{
Function checks whether given numbers are coprime.
}
