% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dlog.ph}
\alias{dlog.ph}
\title{Pohling-Hellman discrete logarithm}
\usage{
dlog.ph(a, g, m, disp = TRUE)
}
\arguments{
\item{a}{power}

\item{g}{base}

\item{m}{modular}

\item{disp}{choose whether to see stepwise solution, def. FALSE}
}
\value{
number
}
\description{
Algorith for discrete logarithm, allows user to see stepwise solution
of log g  a   in Zm
}
