% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dlog.shanks}
\alias{dlog.shanks}
\title{Shanks' algorithm for discrete logarithm}
\usage{
dlog.shanks(a, g, m, disp = TRUE, warnings = TRUE)
}
\arguments{
\item{a}{power}

\item{g}{base - prime root preferably}

\item{m}{modulus}

\item{disp}{choose whether to see stepwise solution}

\item{warnings}{choose whether to see warnings}
}
\value{
number
}
\description{
Looks for the discrete logarithm. Base is g - the prime root - preferably; a is power, m modulus.
There might be warning passed, for ex. if g is not a prime root.
Also user can see stepwise version
}
