% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{elliptic.add}
\alias{elliptic.add}
\title{Addition on elliptic curve}
\usage{
elliptic.add(P, Q, a, p, disp = TRUE)
}
\arguments{
\item{P, Q}{points on the curve}

\item{a}{C = x^3 + ax}

\item{p}{modulus}

\item{choose}{whether to see stepwise solution, def. TRUE}
}
\value{
coordinates
}
\description{
Adds two points on curve C = x^3 + ax mod p
}
