% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{elliptic.multiplicity}
\alias{elliptic.multiplicity}
\title{Multiply point on the curve}
\usage{
elliptic.multiplicity(P, k, a, p)
}
\arguments{
\item{P}{point}

\item{k}{power}

\item{a}{C = x^3 + ax}

\item{p}{modulus}

\item{disp}{choose whether to see stepwise solution}
}
\value{
coordinates
}
\description{
Allows to count n*P on the elliptic curve C = x^3 + ax modulo p
Not recommended for large numbers, check then elliptic.mutiplicity.a
}
