% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{elliptic.ord}
\alias{elliptic.ord}
\title{Order of point on elliptic curve}
\usage{
elliptic.ord(P, a, p)
}
\arguments{
\item{P}{point}

\item{a}{C = x^3 + ax}

\item{p}{modulus}
}
\value{
number
}
\description{
Check the order of the point on the elliptic cruve C = x^3 + ax.
}
