% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{euclidean.div}
\alias{euclidean.div}
\title{Euclidean division}
\usage{
euclidean.div(a, b)
}
\arguments{
\item{a}{divident}

\item{b}{divisor}
}
\value{
two-element vector. First quotient, next reminder
}
\description{
Performs Euclidean division on given entries
}
