% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{GCD.ext}
\alias{GCD.ext}
\title{Extended GCD}
\usage{
GCD.ext(a, b, disp = FALSE)
}
\arguments{
\item{a, b}{given entries}

\item{disp}{choose whether to see partial results. Def. FALSE}
}
\value{
vector consisting of GCD followed by x, y such that ax + by = 0
}
\description{
Extended algorithm looking for the greatest common divisor.
Finds solution for the equation ax + by = 0
Allows user to see results step-by-step.
}
