% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{jacobi.symbol}
\alias{jacobi.symbol}
\title{Jacobi symbol}
\usage{
jacobi.symbol(a, n, disp = FALSE)
}
\arguments{
\item{a}{number on the top}

\item{n}{number on the bottom}

\item{disp}{choose whether to display stepwise, def. FALSE}
}
\value{
-1, 0 or 1
}
\description{
Calculates Jacobi symbol, of course also Legendre's symbol.
}
