% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mod.equations}
\alias{mod.equations}
\title{Solves system of simple modular equations}
\usage{
mod.equations(y, m, disp = FALSE)
}
\arguments{
\item{y}{vector of values}

\item{m}{vector of modulus}

\item{disp}{choose whether to see stepwise solutions. By def. FALSE}
}
\value{
data frame
}
\description{
Solves system of equations given by the chinese reminder theorem.
}
