% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mod.inv}
\alias{mod.inv}
\title{Modulus inversion}
\usage{
mod.inv(a, n)
}
\arguments{
\item{a}{will be calcualted a^-1}

\item{n}{modulus}
}
\value{
inverse of a modulo n
}
\description{
Find modulo inversion of given number.
}
