% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mod.power}
\alias{mod.power}
\title{Power funcion in terms of modulus calculations}
\usage{
mod.power(x, k, n, disp = FALSE)
}
\arguments{
\item{x}{number to be powered}

\item{k}{power}

\item{n}{modulus}

\item{disp}{choose whether to see stepwise results. Def. FALSE}
}
\value{
number, the result
}
\description{
Allows user to find great powers of numbers. Here is used the algorithm
of so-called fast powering. User can see stepwise algorithm's results, by default FALSE.
}
