% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mod}
\alias{mod}
\title{Calculate modulus value}
\usage{
mod(a, n)
}
\arguments{
\item{a}{number to be calculated}

\item{n}{modulus}
}
\value{

}
\description{
Equivalent of \%\%, however much more transparent
For given equation x = a mod n
}
