% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ord}
\alias{ord}
\title{Order modulo}
\usage{
ord(x, n, disp = FALSE)
}
\arguments{
\item{x}{element which order is yet to be calculated}

\item{n}{modulus}

\item{disp}{displays additional information if number is the prime root. By default set to FALSE}
}
\value{
order of the element
}
\description{
Function calculates order of the element in cyclic group Zn
}
