% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{power.rem}
\alias{power.rem}
\title{Power remainder}
\usage{
power.rem(a, n, m, disp = TRUE)
}
\arguments{
\item{a}{equivalent value, alike power in discrete logarithm}

\item{n}{power}

\item{m}{modulus}

\item{disp}{choose whether to display stepwise, def. FALSE}
}
\value{
vector: result, modulo
}
\description{
Checks if solution for equation x^n = a mod m exists with respect to x.
}
