% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prime.factors}
\alias{prime.factors}
\title{Factorise number}
\usage{
prime.factors(n)
}
\arguments{
\item{n}{number to be factorised}
}
\value{
data frame
}
\description{
Factorises given number giving prime factors. First it shows number
of the prime factors, next sequence prime factor followed by its multiplicity
}
