% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{printt}
\alias{printt}
\title{Print pasted}
\usage{
printt(x, ...)
}
\arguments{
\item{x}{elements, separated with coma}
}
\value{
pasted text
}
\description{
The function allows user to print pasted, using paste0, text with no nead of writing paste.
}
