% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sq.root.ts}
\alias{sq.root.ts}
\title{Discrete Square Root}
\usage{
sq.root.ts(a, p, disp = TRUE)
}
\arguments{
\item{a}{number to be square rooted}

\item{p}{modulus}
}
\value{
vector consisting of 2 results
}
\description{
Tonell-Shank's algorithm for discrete square root.
}
